# app/web/models/category.go
package models

import "time"

type Category struct {
	ID          string    `json:"id" gorm:"not null;uniqueIndex;primary_key"`
	StoreID     string    `json:"store_id" gorm:"not null;index"`
	Store       Store     `json:"-"`
	BillboardID string    `json:"billboard_id" gorm:"not null;index"`
	Billboard   Billboard `gorm:"foreignKey:BillboardID" json:"billboard"`
	Products    []Product `gorm:"foreignKey:CategoryID" json:"products"`
	Name        string    `json:"name"`
	CreatedAt   time.Time `json:"created_at" gorm:"autoCreateTime"`
	UpdatedAt   time.Time `json:"updated_at" gorm:"autoUpdateTime"`
}

type CategoryResponse struct {
	ID          string    `json:"id"`
	StoreID     string    `json:"store_id"`
	BillboardID string    `json:"billboard_id"`
	Name        string    `json:"name"`
	Products    []Product `json:"products"`
	CreatedAt   time.Time `json:"created_at"`
	UpdatedAt   time.Time `json:"updated_at"`
}

type CategoryCreate struct {
	BillboardID string `json:"billboard_id" validate:"required"`
	Name        string `json:"name" validate:"required,min=4,max=50"`
}

type CategoryUpdate struct {
	BillboardID string `json:"billboard_id" validate:"required"`
	Name        string `json:"name" validate:"required,min=4,max=50"`
}

func ToCategoryResponse(category Category) CategoryResponse {
	return CategoryResponse{
		ID:          category.ID,
		StoreID:     category.StoreID,
		BillboardID: category.BillboardID,
		Name:        category.Name,
		Products:    category.Products,
		CreatedAt:   category.CreatedAt,
		UpdatedAt:   category.UpdatedAt,
	}
}

func ToCategoryResponses(categories []Category) []CategoryResponse {
	var responses []CategoryResponse
	for _, category := range categories {
		responses = append(responses, ToCategoryResponse(category))
	}
	return responses
}